backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
sqlmodel==0.0.8
pydantic==1.10.7
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
email-validator==2.0.0 